cleanup

+++ php/helper/addUserToJSON.php
@@ -12,43 +12,64 @@ if($_SESSION['loggedIn'] == "admin" && $_SERVER['REQUEST_METHOD'] == 'POST')
     $password = $_POST['password'];     //encrypted
     $forbiddenProjects = $_POST['forbiddenProjects'];   //as json --> encoded by javascript
     $accountType = $_POST['accountType'];
+    $alreadyExists = false;
 
-    if (file_exists($pathString))
+    if($username != "admin" && $username != "public")
     {
-        $isFileThere = true;
-    }
-
-    try
-    {
-        if ($isFileThere)
+        //check if username already exists!
+        global $userArray;
+        include "getUsersFromJSON.php";
+        for ($i = 0; $i < sizeof($userArray); $i++)
         {
-            $configFile = file_get_contents($pathString);
-            $array = json_decode($configFile, false);
-
+            if ($userArray[$i]->{'username'} == $username)
+            {
+                $alreadyExists = true;
+                break;
+            }
         }
 
-        $newUserArray = array
-        (
-            "username" => $username,
-            "password" => crypt($password, '$5$g3t#~34uö@$'),
-            "accountType" => $accountType,
-            "forbiddenProjects" => $forbiddenProjects
-        );
+        if(!$alreadyExists)
+        {
+            if (file_exists($pathString))
+            {
+                $isFileThere = true;
+            }
+            try
+            {
+                if ($isFileThere)
+                {
+                    $configFile = file_get_contents($pathString);
+                    $array = json_decode($configFile, false);
 
-        //checks boolean value to see if file is there. If not generates it
-        if ($isFileThere)
-            array_push($array, $newUserArray);
-        else
-            $array[0] = $newUserArray;
+                }
 
-        //push new reg id into array
-        $fileToSave = json_encode($array);
-        file_put_contents($pathString, $fileToSave);
+                $newUserArray = array
+                (
+                    "username" => $username,
+                    "password" => crypt($password, '$5$g3t#~34uö@$'),
+                    "accountType" => $accountType,
+                    "forbiddenProjects" => $forbiddenProjects
+                );
 
+                //checks boolean value to see if file is there. If not generates it
+                if ($isFileThere)
+                    array_push($array, $newUserArray);
+                else
+                    $array[0] = $newUserArray;
+
+                //push new reg id into array
+                $fileToSave = json_encode($array);
+                file_put_contents($pathString, $fileToSave);
+
+            }
+            catch (Exception $e)
+            {
+                var_dump($e);   //TODO remove this from production code sometime
+            }
+
+        }
     }
-    catch (Exception $e)
-    {
-        var_dump($e);   //TODO remove this from production code sometime
-    }
+
+    echo $returnValue;
 
 }
